[ugc] Fake UGC are generated in API now.
