+++ ugc/ugc_tests/CMakeLists.txt
@@ -6,5 +6,5 @@ set(
 )
 
 omim_add_test(${PROJECT_NAME} ${SRC})
-omim_link_libraries(${PROJECT_NAME} platform coding geometry base)
+omim_link_libraries(${PROJECT_NAME} ugc indexer platform coding geometry base)
 link_qt5_core(${PROJECT_NAME})
